added authentication for tags and themes

+++ db_test/restaurantAPI/themes.py
@@ -1,7 +1,8 @@
 from flask import Blueprint, jsonify, request
 from database import query_db, insert_db
 from flasgger import swag_from
-
+from flask_jwt_extended import jwt_required
+import os
 
 themes_blueprint = Blueprint('themes', __name__)
 
@@ -81,6 +82,12 @@ def get_themes():
     }
 })
 def add_theme():
+    auth_header = request.headers.get("Authorization")
+    key = auth_header.split(" ")[1]
+    storedkey = os.getenv("ADMINKEY")
+    if key != storedkey:
+        return jsonify({"message": "Wrong key"}), 400
+    
     data = request.get_json()
     name = data["name"]
     primaryColor = data["primaryColor"]
@@ -129,6 +136,12 @@ def add_theme():
     }
 })
 def delete_theme(themeName):
+    auth_header = request.headers.get("Authorization")
+    key = auth_header.split(" ")[1]
+    storedkey = os.getenv("ADMINKEY")
+    if key != storedkey:
+        return jsonify({"message": "Wrong key"}), 400
+    
     insert_db("""
                 DELETE FROM themes WHERE name = %s
               """,
@@ -138,6 +151,7 @@ def delete_theme(themeName):
 
 
 @themes_blueprint.route("/themes/applyThemeToRestaurant/", methods=["POST"])
+@jwt_required()
 @swag_from({
     'tags': ['Themes'],
     'summary': 'Assign a theme to a restaurant',
